move the current processes in a dedicated redux state
