+++ web/src/js/__tests__/ducks/tutils.ts
@@ -142,21 +142,6 @@ export const testState: RootState = {
             {
                 active: false,
                 selectedProcesses: "",
-                currentProcesses: [
-                    {
-                        is_visible: true,
-                        executable: "curl.exe",
-                        is_system: "false",
-                        display_name: "curl",
-                    },
-                    {
-                        is_visible: true,
-                        executable: "http.exe",
-                        is_system: "false",
-                        display_name: "http",
-                    },
-                ],
-                isLoading: false,
             },
         ],
         wireguard: [
@@ -193,6 +178,23 @@ export const testState: RootState = {
                 active: false,
             },
         ],
+    },
+    processes: {
+        currentProcesses: [
+            {
+                is_visible: true,
+                executable: "curl.exe",
+                is_system: false,
+                display_name: "curl",
+            },
+            {
+                is_visible: true,
+                executable: "http.exe",
+                is_system: false,
+                display_name: "http",
+            },
+        ],
+        isLoading: false,
     },
 };
