+++ web/src/js/modes/local.ts
@@ -1,15 +1,6 @@
 import { ModeState, RawSpecParts } from ".";
 
-export interface Process {
-    is_visible: boolean;
-    executable: string;
-    is_system: string;
-    display_name: string;
-}
-
 export interface LocalState extends ModeState {
-    isLoading: boolean;
-    currentProcesses: Process[];
     selectedProcesses?: string;
 }
 
@@ -20,7 +11,5 @@ export const getSpec = (m: LocalState): string => {
 export const parseRaw = ({ data }: RawSpecParts): LocalState => ({
     ui_id: Math.random(),
     active: true,
-    isLoading: false,
-    currentProcesses: [],
     selectedProcesses: data,
 });
